calculation_service_keycloak/requirements.txt
httpx
fastapi
uvicorn
python-jose[cryptography]